refactor(testcolor): share the ANSI test helpers across the workspace
gsw kept `with_forced_ansi` and `strip_ansi` as private test-only items of its
own binary crate, so no other member could reach them. cwt then shipped two
render tests that compare colored output against plain text. Such a test passes
when the run writes to a file. It fails when the run writes to a terminal, which
is what a hand-typed `git commit` gives it.

Move both helpers, and `max_red_channel` and `TRUECOLOR_FG` beside them, into a
new `testcolor` library crate. Any member takes it as a dev-dependency.
`strip_ansi` arrives with the direct tests it never had, including a case for
multi-byte glyphs.

Move gsw's clippy.toml to the workspace root, so the ban on
`colored::control::set_override` covers every member instead of gsw alone.
seescc calls the raw API in production. It now says so at the call site with a
reason, where a reader sees it.

No behavior changes. The moved code and its tests are unchanged apart from
visibility.

Mutation check of the moved ban: a `colored::control::set_override` call added
to cwt made `cargo clippy -p cwt --all-targets` fail with "use of a disallowed
method". Removing the call restored green. cwt carried no clippy.toml before
this commit, so the same call passed clippy then.
